+++ app/Http/Controllers/User/Dashboard/BlogController.php
@@ -11,16 +11,19 @@ class BlogController extends Controller
 {
     public function index()
     {
+        $this->authorize('read blog');
         $blogs = Blog::dynamicPaginate();
         return view('dashboard.blogs.index', compact('blogs'));
     }
     public function create()
     {
+        $this->authorize('create blog');
         return view('dashboard.blogs.create');
     }
 
     public function store(CreateBlogRequest $request)
     {
+        $this->authorize('create blog');
         $blog = request()->user()->blogs()->create($request->only(['title', 'context']));
         $blog->addMediaFromRequest('image')->ToMediaCollection('blog_images');
 
@@ -29,12 +32,14 @@ class BlogController extends Controller
 
     public function edit(Blog $blog)
     {
+        $this->authorize('edit blog');
         $blog->load('author');
         return view('dashboard.blogs.edit', compact('blog'));
     }
 
     public function update(UpdateBlogRequest $request, Blog $blog)
     {
+        $this->authorize('update blog');
         $blog->update($request->validated());
         if($request->hasFile('image')){
             $blog->clearMediaCollection();
@@ -46,12 +51,14 @@ class BlogController extends Controller
 
     public function show(Blog $blog)
     {
+        $this->authorize('read blog');
         $blog->load('author');
         return view('dashboard.blogs.show', compact('blog'));
     }
 
     public function destroy(Blog $blog)
     {
+        $this->authorize('delete blog');
         $blog->account()->delete();
 
         return redirect()->route('user.dashboard.blogs.index')->with('success', 'Blog deleted successfully.');
