+++ app/Http/Controllers/User/Dashboard/DiscountController.php
@@ -11,6 +11,7 @@ class DiscountController extends Controller
 {
     public function index()
     {
+        $this->authorize('read discount');
         $active = Discount::isActive()->sum('amount');
         $not_active = Discount::isNotActive()->sum('amount');
         $expired = Discount::isExpired()->sum('amount');
@@ -42,11 +43,13 @@ class DiscountController extends Controller
 
     public function create()
     {
+        $this->authorize('create discount');
         return view('user.dashboard.reports.discounts.create');
     }
 
     public function store(CreateDiscountRequest $request)
     {
+        $this->authorize('create discount');
         $discount = Discount::create($request->validated());
 
         return to_route('user.dashboard.reports.discounts.index')->with('success', 'created successfully');
@@ -54,16 +57,19 @@ class DiscountController extends Controller
 
     public function show(Discount $discount)
     {
+        $this->authorize('read discount');
         return view('user.dashboard.reports.discounts.show', compact('discount'));
     }
 
     public function edit(Discount $discount)
     {
+        $this->authorize('update discount');
         return view('user.dashboard.reports.discounts.edit', compact('discount'));
     }
 
     public function update(UpdateDiscountRequest $request, Discount $discount)
     {
+        $this->authorize('update discount');
         $discount->update($request->validated());
 
         return to_route('user.dashboard.reports.discounts.index')->with('success', 'updated successfully');
@@ -71,6 +77,7 @@ class DiscountController extends Controller
 
     public function destroy(Discount $discount)
     {
+        $this->authorize('delete discount');
         $discount->delete();
 
         return to_route('user.dashboard.reports.discounts.index')->with('success', 'deleted successfully');
