+++ app/Http/Controllers/User/Dashboard/InvoiceController.php
@@ -11,6 +11,7 @@ class InvoiceController extends Controller
 {
     public function index()
     {
+        $this->authorize('read invoice');
         $paid = Invoice::isPaid()->sum('amount');
         $new = Invoice::isnew()->sum('amount');
         $overdue = Invoice::isOverdue()->sum('amount');
@@ -38,11 +39,13 @@ class InvoiceController extends Controller
 
     public function create()
     {
+        $this->authorize('create invoice');
         return view('user.dashboard.reports.invoices.create');
     }
 
     public function store(CreateInvoiceRequest $request)
     {
+        $this->authorize('create invoice');
         $invoice = Invoice::create($request->validated());
 
         return to_route('user.dashboard.reports.invoices.index')->with('success', 'created successfully');
@@ -50,16 +53,19 @@ class InvoiceController extends Controller
 
     public function show(Invoice $invoice)
     {
+        $this->authorize('read invoice');
         return view('user.dashboard.reports.invoices.show', compact('invoice'));
     }
 
     public function edit(Invoice $invoice)
     {
+        $this->authorize('update invoice');
         return view('user.dashboard.reports.invoices.edit', compact('invoice'));
     }
 
     public function update(UpdateInvoiceRequest $request, Invoice $invoice)
     {
+        $this->authorize('update invoice');
         $invoice->update($request->validated());
 
         return to_route('user.dashboard.reports.invoices.index')->with('success', 'updated successfully');
@@ -67,6 +73,7 @@ class InvoiceController extends Controller
 
     public function destroy(Invoice $invoice)
     {
+        $this->authorize('delete invoice');
         $invoice->delete();
 
         return to_route('user.dashboard.reports.invoices.index')->with('success', 'deleted successfully');
