+++ app/Http/Controllers/User/Dashboard/WorkerController.php
@@ -11,33 +11,39 @@ class WorkerController extends Controller
 {
     public function index()
     {
+        $this->authorize('read worker');
         $workers = Worker::with('office.account')->dynamicPaginate();
         return view('dashboard.workers.index', compact('workers'));
     }
 
     public function create()
     {
+        $this->authorize('create worker');
         return view('dashboard.workers.create');
     }
 
     public function store(CreateWorkerRequest $request)
     {
+        $this->authorize('create worker');
         $worker = Worker::create($request->validated());
         return redirect()->route('user.dashboard.workers.index')->with('success', 'Worker created successfully.');
     }
 
     public function show(Worker $worker)
     {
+        $this->authorize('read worker');
         return view('dashboard.workers.show', compact('worker'));
     }
 
     public function edit(Worker $worker)
     {
+        $this->authorize('update worker');
         return view('dashboard.workers.edit', compact('worker'));
     }
 
     public function update(UpdateWorkerRequest $request, Worker $worker)
     {
+        $this->authorize('update worker');
         $worker->update($request->validated());
         $request->validate(['main_image' => ['sometimes', 'image', 'max:2048']]);
         if ($request->hasFile('main_image')) {
@@ -50,6 +56,7 @@ class WorkerController extends Controller
 
     public function destroy(Worker $worker)
     {
+        $this->authorize('delete worker');
         $worker->delete();
         return redirect()->route('user.dashboard.workers.index')->with('success', 'Worker deleted successfully.');
     }
